Added GlobalAnalysis Plugin
