+++ gui/main.ui
@@ -153,7 +153,6 @@
     <addaction name="menuImport"/>
     <addaction name="actionSave_Points"/>
     <addaction name="actionSave_Movie"/>
-    <addaction name="actionChange_Internal_Data_type"/>
     <addaction name="actionSettings"/>
    </widget>
    <widget class="QMenu" name="menuImage">
